requirements.txt
pyTelegramBotAPI==4.6.0
requests==2.26.0
python-dotenv==0.20.0
